adk-agent/pyproject.toml
[project]
name = "adk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.53.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "akshare>=1.15.0",
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "edgartools>=0.0.1",
    "tavily-python>=0.3.0",
    "duckduckgo-search>=5.0.0",
    "google-search-results>=2.4.0",
    "llama-cloud>=0.0.1",
    "pymupdf>=1.23.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "litellm>=1.75.5",
]
